requirements.txt
Flask==2.2.5
gunicorn==23.0.0
joblib==1.4.2
transformers==4.42.4
tensorflow-cpu==2.17.0
tf_keras==2.17.0
